sentinel zuul rename package name
